Fix for issues... Issue #49 Issue #50 Issue #58 Issue #57 Issue #55 Issue #48 Issue #46 Issue #40
