+++ Microsoft.Dynamics365.UIAutomation.Api/Pages/XrmEntityPage.cs
@@ -4,7 +4,6 @@
 using Microsoft.Dynamics365.UIAutomation.Browser;
 using OpenQA.Selenium;
 using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 
@@ -226,145 +225,6 @@ namespace Microsoft.Dynamics365.UIAutomation.Api
 
                 return true;
             });
-        }
-
-        /// <summary>
-        /// Set Lookup Value for the field
-        /// </summary>
-        /// <param name="field">The Field</param>
-        /// <param name="index">The Index</param>
-        /// <example>xrmBrowser.Entity.SelectLookup("customerid", 0);</example>
-        public BrowserCommandResult<bool> SelectLookup(string field, [Range(0, 9)]int index)
-        {
-            return this.Execute(GetOptions($"Set Lookup Value: {field}"), driver => 
-            {
-                if (driver.HasElement(By.Id(field)))
-                {
-                    var input = driver.ClickWhenAvailable(By.Id(field));
-
-                    if (input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])) == null)
-                        throw new InvalidOperationException($"Field: {field} is not lookup");
-
-                    input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])).Click();
-
-                    var dialogName = $"Dialog_{field}_IMenu";
-                    var dialog = driver.FindElement(By.Id(dialogName));
-
-                    var dialogItems = OpenDialog(dialog).Value;
-
-                    if (dialogItems.Count < index)
-                        throw new InvalidOperationException($"List does not have {index + 1} items.");
-
-                    var dialogItem = dialogItems[index];
-                    dialogItem.Element.Click();
-                }
-                else
-                    throw new InvalidOperationException($"Field: {field} Does not exist");
-
-                return true;
-            });
-        }
-
-        /// <summary>
-        /// Set Lookup Value for the field
-        /// </summary>
-        /// <param name="field">The Field</param>
-        /// <param name="value">The Lookup value</param>
-        public BrowserCommandResult<bool> SelectLookup(string field, string value)
-        {
-            return this.Execute(GetOptions($"Set Lookup Value: {field}"), driver=> 
-            {
-                if (driver.HasElement(By.Id(field)))
-                {
-                    var input = driver.ClickWhenAvailable(By.Id(field));
-
-                    if (input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])) == null)
-                        throw new InvalidOperationException($"Field: {field} is not lookup");
-
-                    var lookupIcon = input.FindElement(By.ClassName("Lookup_RenderButton_td"));
-                    lookupIcon.Click();
-
-                    var dialogName = $"Dialog_{field}_IMenu";
-                    var dialog = driver.FindElement(By.Id(dialogName));
-
-                    var dialogItems = OpenDialog(dialog).Value;
-
-                    if (!dialogItems.Exists(x => x.Title == value))
-                        throw new InvalidOperationException($"List does not have {value}.");
-
-                    var dialogItem = dialogItems.Where(x => x.Title == value).First();
-                    dialogItem.Element.Click();
-                }
-                else
-                    throw new InvalidOperationException($"Field: {field} Does not exist");
-
-                return true;
-            });
-        }
-
-        /// <summary>
-        /// Set Lookup Value for the field
-        /// </summary>
-        /// <param name="field">The Field</param>
-        /// <param name="openLookupPage">The Open Lookup Page</param>
-        public BrowserCommandResult<bool> SelectLookup(string field, bool openLookupPage = true)
-        {
-            return this.Execute(GetOptions($"Set Lookup Value: {field}"), driver=> 
-            {
-                if (driver.HasElement(By.Id(field)))
-                {
-                    var input = driver.ClickWhenAvailable(By.Id(field));
-
-                    if (input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])) == null)
-                        throw new InvalidOperationException($"Field: {field} is not lookup");
-
-                    input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])).Click();
-
-                    var dialogName = $"Dialog_{field}_IMenu";
-                    var dialog = driver.FindElement(By.Id(dialogName));
-
-                    var dialogItems = OpenDialog(dialog).Value;
-
-                    if (dialogItems.Any())
-                    {
-                        var dialogItem = dialogItems.Last();
-                        dialogItem.Element.Click();
-                    }
-                }
-                else
-                    throw new InvalidOperationException($"Field: {field} Does not exist");
-
-                return true;
-            });
-        }
-  
-        /// <summary>
-        /// Opens the dialog
-        /// </summary>
-        /// <param name="dialog"></param>
-        private BrowserCommandResult<List<XrmListItem>> OpenDialog(IWebElement dialog)
-        {
-            var list = new List<XrmListItem>();
-            var dialogItems = dialog.FindElements(By.TagName("li"));
-
-            foreach (var dialogItem in dialogItems)
-            {
-                if (dialogItem.GetAttribute("role") != null && dialogItem.GetAttribute("role") == "menuitem")
-                {
-                    var links = dialogItem.FindElements(By.TagName("a"));
-
-                    if (links != null && links.Count > 1)
-                    {
-                        var title = links[1].GetAttribute("title");
-
-                        list.Add(new XrmListItem() {
-                            Title = title,
-                            Element = links[1] });
-                    }
-                }
-            }
-
-            return list;
         }
 
         /// <summary>
