+++ Microsoft.Dynamics365.UIAutomation.Api/Pages/XrmGridPage.cs
@@ -124,107 +124,6 @@ namespace Microsoft.Dynamics365.UIAutomation.Api
             });
         }
 
-        /// <summary>
-        /// Refreshes this instance.
-        /// </summary>
-        /// <example></example>
-        public BrowserCommandResult<bool> Refresh(int thinkTime = Constants.DefaultThinkTime)
-        {
-            Browser.ThinkTime(thinkTime);
-
-            return this.Execute(GetOptions("Refresh"), driver =>
-            {
-                driver.ClickWhenAvailable(By.XPath(Elements.Xpath[Reference.Grid.Refresh]));
-
-                return true;
-            });
-        }
-
-        /// <summary>
-        /// Firsts the page.
-        /// </summary>
-        /// <param name="thinkTime">Used to simulate a wait time between human interactions. The Default is 2 seconds.</param>
-        /// <example>xrmBrowser.Grid.FirstPage();</example>
-        public BrowserCommandResult<bool> FirstPage(int thinkTime = Constants.DefaultThinkTime)
-        {
-            Browser.ThinkTime(thinkTime);
-
-            return this.Execute(GetOptions("FirstPage"), driver =>
-            {
-                var firstPageIcon = driver.FindElement(By.XPath(Elements.Xpath[Reference.Grid.FirstPage]));
-
-                if (firstPageIcon.GetAttribute("disabled") != null)
-                    return false;
-                else
-                    firstPageIcon.Click();
-                return true;
-            });
-        }
-
-        /// <summary>
-        /// Nexts the page.
-        /// </summary>
-        /// <param name="thinkTime">Used to simulate a wait time between human interactions. The Default is 2 seconds.</param>
-        /// <example>xrmBrowser.Grid.NextPage();</example>
-        public BrowserCommandResult<bool> NextPage(int thinkTime = Constants.DefaultThinkTime)
-        {
-            Browser.ThinkTime(thinkTime);
-
-            return this.Execute(GetOptions("Next"), driver =>
-            {
-                var nextIcon = driver.FindElement(By.XPath(Elements.Xpath[Reference.Grid.NextPage]));
-
-                if (nextIcon.GetAttribute("disabled") != null)
-                    return false;
-                else
-                    nextIcon.Click();
-                return true;
-            });
-        }
-
-        /// <summary>
-        /// Toggles the select all.
-        /// </summary>
-        /// <param name="thinkTime">Used to simulate a wait time between human interactions. The Default is 2 seconds.</param>
-        /// <example>xrmBrowser.Grid.SelectAllRecords();</example>
-        public BrowserCommandResult<bool> SelectAllRecords(int thinkTime = Constants.DefaultThinkTime)
-        {
-            Browser.ThinkTime(thinkTime);
-
-            return this.Execute(GetOptions("ToggleSelectAll"), driver =>
-            {
-                // We can check if any record selected by using
-                // driver.FindElements(By.ClassName("ms-crm-List-SelectedRow")).Count == 0
-                // but this function doesn't check it.
-                var selectAll = driver.WaitUntilAvailable(By.XPath(Elements.Xpath[Reference.Grid.ToggleSelectAll]),
-                          "The Toggle SelectAll is not available.");
-
-                selectAll.Click();
-
-                return true;
-            });
-        }
-
-        /// <summary>
-        /// Previouses the page.
-        /// </summary>
-        /// <param name="thinkTime">Used to simulate a wait time between human interactions. The Default is 2 seconds.</param>
-        /// <example>xrmBrowser.Grid.PreviousPage();</example>
-        public BrowserCommandResult<bool> PreviousPage(int thinkTime = Constants.DefaultThinkTime)
-        {
-            Browser.ThinkTime(thinkTime);
-
-            return this.Execute(GetOptions("PreviousPage"), driver =>
-            {
-                var previousIcon = driver.FindElement(By.XPath(Elements.Xpath[Reference.Grid.PreviousPage]));
-
-                if (previousIcon.GetAttribute("disabled") != null)
-                    return false;
-                else
-                    previousIcon.Click();
-                return true;
-            });
-        }
 
         /// <summary>
         /// Opens the chart.
@@ -316,61 +215,6 @@ namespace Microsoft.Dynamics365.UIAutomation.Api
                 else
                     sortCol.Click();
                 return true;
-            });
-        }
-
-        /// <summary>
-        /// Gets the grid items.
-        /// </summary>
-        /// <example>xrmBrowser.Grid.GetGridItems();</example>
-        public BrowserCommandResult<List<XrmGridItem>> GetGridItems()
-        {
-            return this.Execute(GetOptions("Get Grid Items"), driver =>
-            {
-                var returnList = new List<XrmGridItem>();
-
-                var itemsTable = driver.FindElement(By.XPath(@"//*[@id=""gridBodyTable""]/tbody"));
-                var columnGroup = driver.FindElement(By.XPath(@"//*[@id=""gridBodyTable""]/colgroup"));
-
-                var rows = itemsTable.FindElements(By.TagName("tr"));
-
-                foreach (var row in rows)
-                {
-                    if (!string.IsNullOrEmpty(row.GetAttribute("oid")))
-                    {
-                        Guid id = Guid.Parse(row.GetAttribute("oid"));
-                        var link =
-                            $"{new Uri(driver.Url).Scheme}://{new Uri(driver.Url).Authority}/main.aspx?etn={row.GetAttribute("otypename")}&pagetype=entityrecord&id=%7B{id:D}%7D";
-
-                        var item = new XrmGridItem
-                        {
-                            EntityName = row.GetAttribute("otypename"),
-                            Id = id,
-                            Url = new Uri(link)
-                        };
-
-                        var cells = row.FindElements(By.TagName("td"));
-                        var idx = 0;
-
-                        foreach (var column in columnGroup.FindElements(By.TagName("col")))
-                        {
-                            var name = column.GetAttribute<string>("name");
-
-                            if (!string.IsNullOrEmpty(name)
-                                && column.GetAttribute("class").Contains(Elements.CssClass[Reference.Grid.DataColumn])
-                                && cells.Count > idx)
-                            {
-                                item[name] = cells[idx].Text;
-                            }
-
-                            idx++;
-                        }
-
-                        returnList.Add(item);
-                    }
-                }
-
-                return returnList;
             });
         }
