+++ Microsoft.Dynamics365.UIAutomation.Api/Pages/XrmQuickCreatePage.cs
@@ -4,8 +4,6 @@
 using Microsoft.Dynamics365.UIAutomation.Browser;
 using OpenQA.Selenium;
 using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq;
 
 namespace Microsoft.Dynamics365.UIAutomation.Api
@@ -61,144 +59,6 @@ namespace Microsoft.Dynamics365.UIAutomation.Api
                 driver.FindElement(By.XPath(Elements.Xpath[Reference.QuickCreate.Save]))?.Click();
                 return true;
             });
-        }
-
-        /// <summary>
-        /// Set Lookup Value for the field
-        /// </summary>
-        /// <param name="field">The Field Value</param>
-        /// <param name="index">The Index</param>
-        /// <example>xrmBrowser.QuickCreate.SelectLookup("primarycontactid", 0);</example>
-        public BrowserCommandResult<bool> SelectLookup(string field, [Range(0, 9)]int index)
-        {
-            return this.Execute(GetOptions($"Set Lookup Value: {field}"), driver =>
-            {
-                if (driver.HasElement(By.Id(field)))
-                {
-                    var input = driver.ClickWhenAvailable(By.Id(field));
-
-                    if (input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])) == null)
-                        throw new InvalidOperationException($"Field: {field} is not lookup");
-
-                    input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])).Click();
-
-                    var dialogName = $"Dialog_{field}_IMenu";
-                    var dialog = driver.FindElement(By.Id(dialogName));
-
-                    var dialogItems = OpenDialog(dialog).Value;
-
-                    if (dialogItems.Count < index)
-                        throw new InvalidOperationException($"List does not have {index + 1} items.");
-
-                    var dialogItem = dialogItems.Values.ToList()[index];
-                    dialogItem.Click();
-                }
-                else
-                    throw new InvalidOperationException($"Field: {field} Does not exist");
-
-                return true;
-            });
-        }
-
-        /// <summary>
-        /// Set Lookup Value for the field
-        /// </summary>
-        /// <param name="field">The Field</param>
-        /// <param name="value">The Value</param>
-        public BrowserCommandResult<bool> SelectLookup(string field, string value)
-        {
-            return this.Execute(GetOptions($"Set Lookup Value: {field}"), driver =>
-            {
-                if (driver.HasElement(By.Id(field)))
-                {
-                    var input = driver.ClickWhenAvailable(By.Id(field));
-
-                    if (input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])) == null)
-                        throw new InvalidOperationException($"Field: {field} is not lookup");
-
-                    var lookupIcon = input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass]));
-                    lookupIcon.Click();
-
-                    var dialogName = $"Dialog_{field}_IMenu";
-                    var dialog = driver.FindElement(By.Id(dialogName));
-
-                    var dialogItems = OpenDialog(dialog).Value;
-
-                    if (!dialogItems.Keys.Contains(value))
-                        throw new InvalidOperationException($"List does not have {value}.");
-
-                    var dialogItem = dialogItems[value];
-                    dialogItem.Click();
-                }
-                else
-                    throw new InvalidOperationException($"Field: {field} Does not exist");
-
-                return true;
-            });
-        }
-
-        /// <summary>
-        /// Set Lookup Value
-        /// </summary>
-        /// <param name="field">The Field</param>
-        /// <param name="openLookupPage">The Open Lookup Page</param>
-        public BrowserCommandResult<bool> SelectLookup(string field, bool openLookupPage = true)
-        {
-            return this.Execute(GetOptions($"Set Lookup Value: {field}"), driver =>
-            {
-                if (driver.HasElement(By.Id(field)))
-                {
-                    var input = driver.ClickWhenAvailable(By.Id(field));
-
-                    if (input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])) == null)
-                        throw new InvalidOperationException($"Field: {field} is not lookup");
-
-                    input.FindElement(By.ClassName(Elements.CssClass[Reference.SetValue.LookupRenderClass])).Click();
-
-                    var dialogName = $"Dialog_{field}_IMenu";
-                    var dialog = driver.FindElement(By.Id(dialogName));
-
-                    var dialogItems = OpenDialog(dialog).Value;
-
-                    var dialogItem = dialogItems.Values.Last();
-
-                    try
-                    {
-                        dialogItem.Click();
-                    }
-                    catch (OpenQA.Selenium.StaleElementReferenceException)
-                    {
-                        // Expected error
-                    }
-                }
-                else
-                    throw new InvalidOperationException($"Field: {field} Does not exist");
-
-                return true;
-            });
-        }
-
-        private BrowserCommandResult<Dictionary<string, IWebElement>> OpenDialog(IWebElement dialog)
-        {
-            var dictionary = new Dictionary<string, IWebElement>();
-            var dialogItems = dialog.FindElements(By.TagName("li"));
-
-            foreach (var dialogItem in dialogItems)
-            {
-                if (dialogItem.GetAttribute("role") != null && dialogItem.GetAttribute("role") == "menuitem")
-                {
-                    var links = dialogItem.FindElements(By.TagName("a"));
-
-                    if (links != null && links.Count > 1)
-                    {
-                        var title = links[1].GetAttribute("title");
-
-                        dictionary.Add(title, links[1]);
-                    }
-                }
-            }
-
-            return dictionary;
         }
 
         /// <summary>
