+++ Microsoft.Dynamics365.UIAutomation.Browser/BrowserType.cs
@@ -9,7 +9,6 @@ namespace Microsoft.Dynamics365.UIAutomation.Browser
         Chrome = 2,
         Firefox = 4,
         Edge = 8,
-        PhantomJs = 16,
-        Remote = 32
+        Remote = 16
     }
 }
